pyproject.toml
[project]
name = "pia"
version = "0.1.0"
description = "Personal Intelligence Agency (PIA) Core"
dependencies = [
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "sqlalchemy>=2.0.27",
    "geoalchemy2>=0.14.3",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "black>=24.2.0",
    "mypy>=1.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 88
select = ["E", "F", "B", "I"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
